requirements.txt
Django==2.2.5
django-crispy-forms==1.7.2
django-filtered-datatables



